optimization around making dataStewardUser from userClass
